handle network connection problem

+++ app/src/main/java/com/king250/kirafan/model/view/MainState.kt
@@ -9,7 +9,7 @@ import android.content.IntentFilter
 import android.os.Build
 import android.widget.Toast
 import androidx.lifecycle.AndroidViewModel
-import com.alibaba.fastjson2.JSON
+import com.google.gson.JsonParser
 import com.king250.kirafan.Env
 import com.king250.kirafan.model.data.UserItem
 import com.king250.kirafan.util.Utils
@@ -36,12 +36,15 @@ class MainState(application: Application) : AndroidViewModel(application) {
 
     private val handler = object : BroadcastReceiver() {
         override fun onReceive(p0: Context?, p1: Intent?) {
+            val context = getApplication<Application>().applicationContext
             when (p1?.getIntExtra("action", -1)) {
                 Env.SERVICE_STARTED -> {
                     _isConnect.value = true
+                    Toast.makeText(context, "已连接至专网", Toast.LENGTH_SHORT).show()
                 }
                 Env.SERVICE_STOPPED -> {
                     _isConnect.value = false
+                    Toast.makeText(context, "已断开连接", Toast.LENGTH_SHORT).show()
                 }
                 else -> {}
             }
@@ -71,10 +74,12 @@ class MainState(application: Application) : AndroidViewModel(application) {
             if (it.endsWith(".dat")) {
                 val input = assets.open(it)
                 val file = File(dir, it)
-                val output = FileOutputStream(file)
-                input.copyTo(output)
-                input.close()
-                output.close()
+                if (!file.exists()) {
+                    val output = FileOutputStream(file)
+                    input.copyTo(output)
+                    input.close()
+                    output.close()
+                }
             }
         }
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
@@ -86,27 +91,27 @@ class MainState(application: Application) : AndroidViewModel(application) {
     }
 
     suspend fun fetch(token: String) {
-        val context = getApplication<Application>().applicationContext
-        val request = Request.Builder()
-            .url("${Env.QLOGIN_API}/me")
-            .header("Authorization", "Bearer $token")
-            .build()
-        val response = Utils.httpClient.newCall(request).execute()
         try {
-            if (response.code == 200) {
-                val result = JSON.parseObject(response.body?.string())
-                response.close()
-                _user.value = UserItem(
-                    result.getString("name"),
-                    result.getString("avatar")
-                )
+            withContext(Dispatchers.IO) {
+                val request = Request.Builder()
+                    .url("${Env.QLOGIN_API}/me")
+                    .header("Authorization", "Bearer $token")
+                    .build()
+                val response = Utils.httpClient.newCall(request).execute()
+                if (response.code == 200) {
+                    val result = JsonParser.parseString(response.body?.string()).asJsonObject
+                    response.close()
+                    _user.value = UserItem(
+                        result.get("name").asString,
+                        result.get("avatar").asString
+                    )
+                }
             }
         }
         catch (e: Exception) {
+            val context = getApplication<Application>().applicationContext
             e.printStackTrace()
-            withContext(Dispatchers.Main) {
-                Toast.makeText(context, "网络好像不太好~", Toast.LENGTH_LONG).show()
-            }
+            Toast.makeText(context, "网络好像不太好~", Toast.LENGTH_LONG).show()
         }
         _isDisableLogin.value = false
     }
